refactor(launcher): split facade, engine, and runtime packages
