+++ internal/launcher/engine/natives.go
@@ -1,4 +1,4 @@
-package launcher
+package engine
 
 import (
 	"archive/zip"
