+++ internal/launcher/engine/priority_windows.go
@@ -1,6 +1,6 @@
 //go:build windows
 
-package launcher
+package engine
 
 import (
 	"fmt"
