+++ internal/launcher/engine/version_overrides.go
@@ -1,4 +1,4 @@
-package launcher
+package engine
 
 import (
 	"log"
